pyproject.toml
[project]
name = "madr"
version = "0.1.0"
description = "Meu Acervo De Romances"
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
    "alembic>=1.13.3",
    "bcrypt==4.0.1",
    "factory-boy>=3.3.1",
    "fastapi[standard]>=0.115.0",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.9",
    "pydantic-settings>=2.5.2",
    "pydantic>=2.9.2",
    "pyjwt>=2.9.0",
    "setuptools>=75.1.0",
    "sqlalchemy>=2.0.35",
    "testcontainers>=4.8.1",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.3",
    "pytest-cov>=5.0.0",
    "ruff>=0.6.9",
    "taskipy>=1.13.0",
]


[tool.ruff]
line-length = 119
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.taskipy.tasks]
lint = 'ruff check .; ruff check . --diff'
format = 'ruff check . --fix & ruff format .'
run = 'fastapi dev madr/app.py'
pre_test = 'task lint'
test = 'pytest -s --cov=. -vv'
post_test = 'coverage html'
